// src/app/core/api/generated/model/campaignDetail.ts
/**
 * Play&Go Project
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 2.0
 * Contact: info@smartcommunitylab.it
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CampaignDetail {
  content?: string;
  extUrl?: string;
  name?: string;
  type?: CampaignDetail.TypeEnum;
}
export namespace CampaignDetail {
  export type TypeEnum = 'faq' | 'privacy' | 'rewards' | 'rules' | 'sponsor';
  export const TypeEnum = {
    Faq: 'faq' as TypeEnum,
    Privacy: 'privacy' as TypeEnum,
    Rewards: 'rewards' as TypeEnum,
    Rules: 'rules' as TypeEnum,
    Sponsor: 'sponsor' as TypeEnum,
  };
}